pyproject.toml
[project]
name = "scene-classification"
version = "0.1.0"
description = "Object-count features (YOLOv8) vs. transfer-learned CNN on indoor scene classification."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Simon Yosboon" }]
license = { text = "MIT" }
dependencies = [
    "numpy>=1.26",
    "pandas>=2.2",
    "scikit-learn>=1.5",
    "xgboost>=2.1",
    "shap>=0.46",
    "ultralytics>=8.3",
    "torch>=2.4",
    "torchvision>=0.19",
    "pillow>=10.4",
    "matplotlib>=3.9",
    "mlflow>=2.17",
    "typer>=0.12",
    "rich>=13.9",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "python-multipart>=0.0.12",
    "pydantic>=2.9",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-cov>=5.0",
    "ruff>=0.7",
    "httpx>=0.27",
]

[project.scripts]
scenes = "scene_classification.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/scene_classification"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
